Seperate non-API spvc code into private source file
Fixes #793
