+++ BUILD.gn
@@ -126,6 +126,8 @@ component("libshaderc_spvc") {
     "libshaderc_spvc/include/shaderc/spvc.h",
     "libshaderc_spvc/include/shaderc/spvc.hpp",
     "libshaderc_spvc/src/spvc.cc",
+    "libshaderc_spvc/src/spvc_private.cc",
+    "libshaderc_spvc/src/spvc_private.h",
   ]
 
   deps = [
